refactor(mercury): split world_data/tests.rs by test subject (#244)
world_data/tests.rs was 942 lines (over the 700-line cap). Split by
test subject into:

  tests/mod.rs            - shared TEST_KEY, sample fixtures, find_setup_stargate_info_args helper
  tests/map_loaded.rs     - mapLoaded multi-packet build/fragmentation/phase-builder tests
  tests/stats.rs          - level_exp + archetype_stats tables
  tests/entity_encoding.rs- BeingAppearance and onEntityTint wire-encoding tests
  tests/bandolier.rs      - bandolier ammo-slot, active-slot, AmmoTypeId regression guards
  tests/stargates.rs      - setupStargateInfo world/known stargate roundtrip tests

All 16 tests pass; no behavior change.
